Improvement in synphot and AstFilter
synphot - Replacement of  if strcmpi by switch lower().
AstFilter.pop_wl - replacement of cumsum by cumtrapz for the calculation of half_width to be coordinated with the filter normalization.
Add a check whether the filter is normalized.
Add treatment for identical values in the cumtrapz which is not supported by interp1.